+++ airbyte-integrations/connectors/source-mailjet-mail/source_mailjet_mail/source.py
@@ -15,4 +15,4 @@ WARNING: Do not modify this file.
 # Declarative Source
 class SourceMailjetMail(YamlDeclarativeSource):
     def __init__(self):
-        super().__init__(**{"path_to_yaml": "mailjet_mail.yaml"})
+        super().__init__(**{"path_to_yaml": "manifest.yaml"})