+++ airbyte-cdk/python/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.25.0
+Use dpath.util.values method to parse response with nested lists
+
+## 0.24.0
+Use dpath.util.values method to parse response with nested lists
+
+## 0.23.0
+Limiting the number of HTTP requests during a test read
+
 ## 0.22.0
 Surface the resolved manifest in the CDK